did some edits

+++ api/User/user.controllers.js
@@ -15,10 +15,22 @@ exports.fetchUser = async (userId, next) => {
 
 exports.getUser = async (req, res, next) => {
   try {
-    const users = await User.find().select("-__v");
+    const users = await User.find().select("-__v -password");
     return res.status(200).json(users);
   } catch (error) {
     return next(error);
+  }
+};
+
+
+exports.getProfile = async (req, res, next) => {
+  try {
+    const users = await User.find({ _id: req.user._id }).select(
+      "-__v -password"
+    );
+    return res.status(200).json(users);
+  } catch (error) {
+    return next({ status: 400, message: error.message });
   }
 };
 
@@ -46,30 +58,37 @@ exports.signin = async (req, res) => {
 exports.signup = async (req, res, next) => {
   // if (req.user.password !== req.user.confirmpassword) display error message
   try {
-    const { password } = req.body;
-    req.body.password = await passhash(password);
+    if (req.file) {
+      req.body.image = `${req.file.path.replace("\\", "/")}`;
+    }
     const newUser = await User.create(req.body);
     const token = generateToken(newUser);
     res.status(201).json({ token });
   } catch (error) {
-    next(error);
+    return next({ status: 400, message: error.message });
   }
 };
 
 exports.updateUser = async (req, res, next) => {
   try {
+    if(!req.user._id.equals(req.foundUser._id)) return next({ status: 400, message: " You are not permitted to update the user" })
+       if (req.file) {
+      req.body.image = `${req.file.path.replace("\\", "/")}`;
+    }
     await User.findByIdAndUpdate(req.user.id, req.body);
     return res.status(204).end();
   } catch (error) {
-    return next(error);
+    return next({ status: 400, message: error.message });
   }
 };
 
+
 exports.deleteUser = async (req, res, next) => {
   try {
+    if(!req.user._id.equals(req.foundUser._id)) return next({ status: 400, message: "You cannot delete other users " })
     await User.findByIdAndRemove({ _id: req.user.id });
     return res.status(204).end();
   } catch (error) {
-    return next(error);
+    return next({ status: 400, message: error.message });
   }
 };
